python-kafka-test/requirements.txt
kafka-python==2.1.2
faker==37.0.1